Added unit tests for stage11

+++ src/by/itstep/makejt/javastages/stage11/task03/controller/Main.java
@@ -10,10 +10,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        double number;
+        int number;
         do {
             System.out.println("Please input natural number:");
-            number = scanner.nextDouble();
+            number = scanner.nextInt();
         } while (number <= 0);
 
         boolean result = PrimeNumber.isPrimeNumber(number);
