+++ src/by/itstep/makejt/javastages/stage11/task04/model/logic/Factorial.java
@@ -1,9 +1,9 @@
 package by.itstep.makejt.javastages.stage11.task04.model.logic;
 
 public class Factorial {
-    public static double countFactorial(double number) {
+    public static int countFactorial(int number) {
 
-        double factorial = 1;
+        int factorial = 1;
 
         while (number > 0) {
             factorial *= number;
